+++ packages/util/src/object/index.ts
@@ -1,0 +1,5 @@
+// Copyright 2017-2021 @polkadot/util authors & contributors
+// SPDX-License-Identifier: Apache-2.0
+
+export { objectClear } from './clear';
+export { objectSpread } from './spread';